samples/pto-approval-agent/pyproject.toml
[project]
name = "pto-approval-agent"
version = "0.0.12"
description = "UiPath LlamaIndex agent that uses human interaction to validate PTO requests."
authors = [{ name = "John Doe", email = "john.doe@myemail.com" }]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
dependencies = [
    "uipath-llamaindex>=0.0.20",
    "llama-index-llms-openai>=0.2.2"
]
